refactor(tests): #58 Tier 1.2 — split OECD full-catalog tests by concern
tests/test_oecd_full_catalog.py (1,131 LOC, 53 tests, 14 Test* classes) →
tests/oecd/{layers,edge_cases_report,infrastructure}.py.

  test_layers.py             8 classes, 28 tests — 10-layer validation
                                                   tests (Layers 1-9, ex.
                                                   Layer 6) — 531 LOC
  test_edge_cases_report.py  2 classes,  6 tests — edge cases +
                                                   cross-layer aggregate
                                                   report — 294 LOC
  test_infrastructure.py     5 classes, 19 tests — client infrastructure:
                                                   rate limiter, parallelism,
                                                   chunking, large dataset
                                                   — 404 LOC

Each split file mirrors the original's full import surface (verbatim),
keeps the original `sys.path.insert(0, ...src)` shim and
`pytestmark = pytest.mark.integration` declaration, and inlines the
per-class fixture / helper references so each file is self-contained.
Test bodies are byte-equivalent.

Verification: pytest collects 53 tests on branch (same as master baseline).
Suite is integration/network-bound and not run in CI; collection-only
verification matches the existing operational pattern.
